catalog.id set to TINYINT, added some "meta-data" columns to dataset

+++ sql/tables/catalog.sql
@@ -3,8 +3,8 @@
  * loaded into the pipeline database.
  */
 
-CREATE TABLE catalog (
-  id INT AUTO_INCREMENT
+CREATE TABLE catalog 
+  (id TINYINT AUTO_INCREMENT
   ,"name" VARCHAR(50) NOT NULL
   ,fullname VARCHAR(250) NOT NULL
   ,PRIMARY KEY (id)
